Protected certain routes

+++ backend/controllers/goalController.js
@@ -1,9 +1,12 @@
 const asyncHandler = require("express-async-handler");
 
 const Goal = require("../model/goalModel");
+const User=require('../model/userModel')
 
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find();
+  const goals = await Goal.find({
+    user: req.user.id,
+  });
   res.status(200).json(goals);
 });
 
@@ -14,6 +17,7 @@ const setGoal = asyncHandler(async (req, res) => {
   }
   const goal = await Goal.create({
     text: req.body.text,
+    user: req.user.id,
   });
   res.status(200).json(goal);
 });
@@ -24,6 +28,20 @@ const updateGoal = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error("Goal not found");
   }
+
+  const user=await User.findById(req.user.id)
+
+  //Checking if user exits or not
+  if(!user){
+    res.status(401)
+    throw new Error('User not found')
+  }
+  //Making sure the logged in user matched the goal user
+  if(goal.user.toString()!==user.id){
+    res.status(401)
+    throw new Error('User not authorized')
+  }
+
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
   });
@@ -36,6 +54,19 @@ const deleteGoal = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error("Goal not found");
   }
+  const user = await User.findById(req.user.id);
+
+  //Checking if user exits or not
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+  //Making sure the logged in user matched the goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
+  }
+  
   const deletedGoal = await Goal.findByIdAndDelete(req.params.id);
   // await goal.remove();
   res.status(200).json({ id: req.params.id, message: deletedGoal.text });
